changed layout of email

+++ capitasocialmedia/socialmedia/mainapp/views.py
@@ -300,8 +300,9 @@ def create_post(request):
 @login_required
 def delete_post(request, post_id):
     post = get_object_or_404(Post, id=post_id, user=request.user)
-    post.delete()
-    messages.success(request, 'Post deleted successfully!')
+    if post.user==request.user:
+     post.delete()
+     messages.success(request, 'Post deleted successfully!')
     return redirect('profile')
